Refactor: Centralize introspection logic in `BaseIntrospectionContext` and extract utilities
- Introduced `BaseIntrospectionContext` for shared state management and type handling in class and function introspectors.
- Moved reusable functions (e.g., `createTypeId`, `isListLike`) to `reflectionUtils.kt` to eliminate duplication.
- Adjusted build settings to include a new project in coverage reports and raised the minimum coverage threshold.
